# ooder-Nexus/src/main/resources/static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ooderNexus API 文档",
    "description": "ooderNexus P2P AI 能力分发中心 - RESTful API 接口文档\n\n## 模块说明\n\n- **OpenWrt P2P**: P2P网络管理、连接建立、QoS配置\n- **MCP Agent**: MCP代理管理、网络状态、命令执行\n- **Device**: 设备管理、控制、日志\n- **System**: 系统配置、监控、健康检查\n- **Network**: 网络路由、链路、终端管理\n- **SkillCenter**: 技能中心、分享、分组\n- **SkillBridge**: 技能桥接、执行、注册\n- **Sync**: 同步任务管理",
    "version": "2.0.0",
    "contact": {
      "name": "ooder Team",
      "email": "support@ooder.net",
      "url": "https://www.ooder.net"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "本地开发环境"
    }
  ],
...
  "paths": {
...
    "/api/skillbridge/register": {
      "post": {
        "tags": ["SkillBridge"],
        "summary": "注册技能",
        "operationId": "registerSkill",
        "requestBody": {
          "content": {"application/json": {"schema": {"type": "object"}}}
        },
        "responses": {"200": {"description": "成功"}}
      }
    },
...
  },
...
}